Split controller_test
