+++ pkg/controller/BUILD.bazel
@@ -50,7 +50,6 @@ go_test(
     name = "go_default_test",
     size = "small",
     srcs = [
-        "controller_test.go",
         "controller_worker_test.go",
         "service_instance_test.go",
         "spec_processor_test.go",
@@ -59,48 +58,20 @@ go_test(
     race = "on",
     deps = [
         "//:go_default_library",
-        "//examples/sleeper:go_default_library",
-        "//examples/sleeper/pkg/apis/sleeper/v1:go_default_library",
         "//pkg/apis/smith/v1:go_default_library",
-        "//pkg/cleanup:go_default_library",
-        "//pkg/cleanup/types:go_default_library",
-        "//pkg/client:go_default_library",
-        "//pkg/client/clientset_generated/clientset/fake:go_default_library",
-        "//pkg/client/smart:go_default_library",
-        "//pkg/plugin:go_default_library",
-        "//pkg/readychecker:go_default_library",
-        "//pkg/readychecker/types:go_default_library",
         "//pkg/resources/apitypes:go_default_library",
-        "//pkg/speccheck:go_default_library",
-        "//pkg/store:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/graph:go_default_library",
-        "//pkg/util/testing:go_default_library",
-        "//vendor/github.com/ash2k/stager:go_default_library",
         "//vendor/github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1:go_default_library",
-        "//vendor/github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/clientset:go_default_library",
-        "//vendor/github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/clientset/fake:go_default_library",
-        "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/require:go_default_library",
-        "//vendor/go.uber.org/zap:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
-        "//vendor/k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/fake:go_default_library",
-        "//vendor/k8s.io/apiextensions-apiserver/pkg/client/informers/externalversions/apiextensions/v1beta1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/meta:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1/unstructured:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/sets:go_default_library",
-        "//vendor/k8s.io/client-go/dynamic:go_default_library",
-        "//vendor/k8s.io/client-go/kubernetes/fake:go_default_library",
-        "//vendor/k8s.io/client-go/rest:go_default_library",
-        "//vendor/k8s.io/client-go/testing:go_default_library",
         "//vendor/k8s.io/client-go/tools/cache:go_default_library",
-        "//vendor/k8s.io/client-go/util/workqueue:go_default_library",
     ],
 )
