+++ pkg/controller_test/detect_infinite_update_cycles_test.go
@@ -1,0 +1,88 @@
+package controller_test
+
+import (
+	"context"
+	"testing"
+
+	smith_v1 "github.com/atlassian/smith/pkg/apis/smith/v1"
+	"github.com/atlassian/smith/pkg/controller"
+	smith_testing "github.com/atlassian/smith/pkg/util/testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	core_v1 "k8s.io/api/core/v1"
+	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/util/sets"
+	kube_testing "k8s.io/client-go/testing"
+	"k8s.io/client-go/tools/cache"
+)
+
+// Should detect infinite update cycles
+func TestDetectInfiniteUpdateCycles(t *testing.T) {
+	t.Parallel()
+	tc := testCase{
+		mainClientObjects: []runtime.Object{
+			configMapNeedsUpdate(),
+		},
+		bundle: &smith_v1.Bundle{
+			ObjectMeta: meta_v1.ObjectMeta{
+				Name:      bundle1,
+				Namespace: testNamespace,
+				UID:       bundle1uid,
+			},
+			Spec: smith_v1.BundleSpec{
+				Resources: []smith_v1.Resource{
+					{
+						Name: "map-needs-update",
+						Spec: smith_v1.ResourceSpec{
+							Object: &core_v1.ConfigMap{
+								TypeMeta: meta_v1.TypeMeta{
+									Kind:       "ConfigMap",
+									APIVersion: core_v1.SchemeGroupVersion.String(),
+								},
+								ObjectMeta: meta_v1.ObjectMeta{
+									Name: mapNeedsAnUpdate,
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		namespace:       testNamespace,
+		expectedActions: sets.NewString("PUT=/api/v1/namespaces/" + testNamespace + "/configmaps/" + mapNeedsAnUpdate),
+		testHandler: fakeActionHandler{
+			response: map[path]fakeResponse{
+				{
+					method: "PUT",
+					path:   "/api/v1/namespaces/" + testNamespace + "/configmaps/" + mapNeedsAnUpdate,
+				}: configMapNeedsUpdateResponse("invalidBundle", "invalidUid"),
+			},
+		},
+		test: func(t *testing.T, ctx context.Context, cntrlr *controller.BundleController, tc *testCase, prepare func(ctx context.Context)) {
+			prepare(ctx)
+			key, err := cache.MetaNamespaceKeyFunc(tc.bundle)
+			require.NoError(t, err)
+			retriable, err := cntrlr.ProcessKey(tc.logger, key)
+			assert.EqualError(t, err, `error processing resource(s): ["`+mapNeedsAnUpdate+`"]`)
+			assert.False(t, retriable)
+			actions := tc.bundleFake.Actions()
+			require.Len(t, actions, 3)
+			bundleUpdate := actions[2].(kube_testing.UpdateAction)
+			assert.Equal(t, testNamespace, bundleUpdate.GetNamespace())
+			updateBundle := bundleUpdate.GetObject().(*smith_v1.Bundle)
+			smith_testing.AssertCondition(t, updateBundle, smith_v1.BundleReady, smith_v1.ConditionFalse)
+			smith_testing.AssertCondition(t, updateBundle, smith_v1.BundleInProgress, smith_v1.ConditionFalse)
+			smith_testing.AssertCondition(t, updateBundle, smith_v1.BundleError, smith_v1.ConditionTrue)
+
+			smith_testing.AssertResourceCondition(t, updateBundle, mapNeedsAnUpdate, smith_v1.ResourceBlocked, smith_v1.ConditionFalse)
+			smith_testing.AssertResourceCondition(t, updateBundle, mapNeedsAnUpdate, smith_v1.ResourceInProgress, smith_v1.ConditionFalse)
+			smith_testing.AssertResourceCondition(t, updateBundle, mapNeedsAnUpdate, smith_v1.ResourceReady, smith_v1.ConditionFalse)
+			resCond := smith_testing.AssertResourceCondition(t, updateBundle, mapNeedsAnUpdate, smith_v1.ResourceError, smith_v1.ConditionTrue)
+			assert.Equal(t, smith_v1.ResourceReasonTerminalError, resCond.Reason)
+			assert.Equal(t, "specification of the created/updated object does not match the desired spec", resCond.Message)
+		},
+	}
+	tc.run(t)
+}
