+++ pkg/controller_test/no_deletions_while_in_progress_test.go
@@ -1,0 +1,52 @@
+package controller_test
+
+import (
+	"testing"
+
+	smith_v1 "github.com/atlassian/smith/pkg/apis/smith/v1"
+
+	sc_v1b1 "github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
+	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+)
+
+// Should not delete resources while Bundle processing is in progress
+func TestNoDeletionsWhileInProgress(t *testing.T) {
+	t.Parallel()
+	tc := testCase{
+		mainClientObjects: []runtime.Object{
+			configMapNeedsDelete(),
+		},
+		scClientObjects: []runtime.Object{
+			serviceInstance(false, true, false),
+		},
+		bundle: &smith_v1.Bundle{
+			ObjectMeta: meta_v1.ObjectMeta{
+				Name:      bundle1,
+				Namespace: testNamespace,
+				UID:       bundle1uid,
+			},
+			Spec: smith_v1.BundleSpec{
+				Resources: []smith_v1.Resource{
+					{
+						Name: resSi1,
+						Spec: smith_v1.ResourceSpec{
+							Object: &sc_v1b1.ServiceInstance{
+								TypeMeta: meta_v1.TypeMeta{
+									Kind:       "ServiceInstance",
+									APIVersion: sc_v1b1.SchemeGroupVersion.String(),
+								},
+								ObjectMeta: meta_v1.ObjectMeta{
+									Name: si1,
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		namespace:            testNamespace,
+		enableServiceCatalog: true,
+	}
+	tc.run(t)
+}
